Improved test coverage
